minor/requirements.txt
pandas
scikit-learn
numpy
fpdf2
matplotlib
joblib
xgboost
scipy